+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - The Parser class.
+- The Cache class.
 
 ## [0.0.1] - 03-23-2025